Simplify api
