+++ packages/lab/src/inlaid-panel/index.ts
@@ -1,4 +1,1 @@
 export * from "./InlaidPanel";
-export * from "./InlaidPanelContent";
-export * from "./InlaidPanelContext";
-export * from "./InlaidPanelTrigger";
